Move to FunctionalJavascript
