+++ .travis.yml
@@ -4,6 +4,12 @@ php:
 sudo: false
 services:
   - mysql
+addons:
+  chrome: stable
+  apt:
+    packages:
+    - chromium-browser
+    - chromium-chromedriver
 cache:
   directories:
     - $HOME/.composer/cache/files
@@ -28,7 +34,8 @@ install:
   - COMPOSER_MEMORY_LIMIT=-1 composer require civicrm/civicrm-asset-plugin:'~1.1' civicrm/civicrm-{core,packages}:'~5.29' drupal/webform_civicrm *@dev
 before_script:
   - cd web
-  - php -S 127.0.0.1:8080 &
+  - php -S 127.0.0.1:8080 2>&1 &
+  - chromedriver 2>&1 &
 script:
   - ../vendor/bin/phpunit -c core modules/contrib/webform_civicrm
 notifications:
