+++ tests/src/FunctionalJavascript/CiviCrmTestBase.php
@@ -1,15 +1,15 @@
 <?php
 
-namespace Drupal\Tests\webform_civicrm\Functional;
+namespace Drupal\Tests\webform_civicrm\FunctionalJavascript;
 
 use Drupal\Core\Database\Database;
-use Drupal\Tests\BrowserTestBase;
+use Drupal\FunctionalJavascriptTests\WebDriverTestBase;
 
 // Requires patching for civicrm-core.
 // @see https://github.com/civicrm/civicrm-core/pull/18843
 // @see https://lab.civicrm.org/dev/core/-/issues/2140
 // @todo move into civicrm-drupal-8 package.
-abstract class CiviCrmTestBase extends BrowserTestBase {
+abstract class CiviCrmTestBase extends WebDriverTestBase {
 
   protected $defaultTheme = 'classy';
